+++ CHANGELOG.md
@@ -3,6 +3,7 @@
 ## 0.1.0-alpha.8 (2026-09-19)
 
 - feat: add presets functionality and summary to the configuration and control APIs
+- feat: enhance user and group account validation, introduce no_user flag for site configuration
 
 ## 0.1.0-alpha.7 (2026-09-19)